refactor: Extract FDv2 URL builder into a shared helper
