+++ launchdarkly-server-sdk/src/fdv2/mod.rs
@@ -2,4 +2,5 @@ pub mod model;
 mod polling;
 mod protocol;
 mod source;
+mod url;
 mod wire;
